+++ HW_23_EmployeesDatabase/src/main/java/configs/Configs.java
@@ -1,4 +1,4 @@
-package homework;
+package configs;
 
 public class Configs {
     public static String mySqlPropFileName = "mysql.properties";
