+++ HW_23_EmployeesDatabase/src/main/java/models/Department.java
@@ -1,4 +1,4 @@
-package homework;
+package models;
 
 public class Department {
     String id;
