+++ HW_23_EmployeesDatabase/src/test/java/AppTest.java
@@ -1,5 +1,3 @@
-package homework;
-
 import static org.junit.Assert.assertTrue;
 
 import org.junit.Test;
